docs: Add "How this Works" section.
Even as someone who has a strong understanding of `systemd`, I found I
needed to read through many of the source files here to get a basic
understanding of how all the pieces fit together.

To make it faster for other's I've created a brief "tour" of the project
with a new "How this works" section in the README.

I've also clarified that the provided units are not enabled by default
and explained how to additional systemd services to run for Sway.